Extract chainable_redirect into its own module
